==> data/17-data-cleanup.r <==

# analyse_data <- analyse_data %>%
#   select(!c("1", "2", "3", "4", "5", "6", "0", "01", "02", "03", "04", "05", "06", "7", "K1", "K2", "T", "W", "Æ", "Å1"))

rm(alle_stemmer,
   geografi,
   vind_treatment,
   vind_stemmesteder,
   kommunalvalg,
   lang_alle_stemmer,
   lang_gruppe_steder,
   stemmer,
   treatment,
   treatment_list)

rm(alle_nrow,
   gem_kolonner)

rm(kv01_treatment,
   kv05_treatment,
   kv09_treatment,
   kv13_treatment,
   kv17_treatment,
   kv21_treatment)
