==> data/30-plots.r <==
library(tidyverse)

extrafont::loadfonts()

## Indlæser farver og skrifttype
source("data/39-farver.r")

## Map plots
source("data/31-maps.r")

## DID plots
source("data/32-did-plots.r")

## Stemmeplots
source("data/33-stemme-plots.r")
