==> data/99-objekt-tjek.r <==
library(tidyverse)
library(testthat)

##################################
####### NUMERISKE VARIABLE #######
##################################

expect_true(exists("antal_obs_samlet"))

##################################
########### REGRESSION ###########
##################################
# reg_tbl1 <- "hej"
expect_true(exists("gt_reg_tbl"))

##################################
######### VISUALISERING ##########
##################################
